refactored player script and enemy resources
- moved player script logic to player controller
- removed enemy resources
