+++ Players/Scenes/player_elgato.tscn
@@ -1,6 +1,7 @@
-[gd_scene load_steps=35 format=3 uid="uid://dkjcbiyswrxb7"]
+[gd_scene load_steps=36 format=3 uid="uid://dkjcbiyswrxb7"]
 
 [ext_resource type="Script" path="res://Players/Scripts/PlayerElgato.cs" id="1_q6xoq"]
+[ext_resource type="Script" path="res://Components/Scripts/PlayerControllerComponent.cs" id="2_40482"]
 [ext_resource type="Texture2D" uid="uid://lo3w8s4ichdj" path="res://Assets/Players/cat_fingerGun-2.png" id="3_wgayd"]
 [ext_resource type="Script" path="res://Components/Scripts/HealthComponent.cs" id="5_b0fly"]
 [ext_resource type="PackedScene" uid="uid://bec21vtufstsw" path="res://Players/Scenes/weapon_elgato.tscn" id="5_c0mry"]
@@ -230,22 +231,26 @@ size = Vector2(5, 5)
 
 [sub_resource type="RectangleShape2D" id="RectangleShape2D_hwqex"]
 
-[node name="player_elgato" type="CharacterBody2D" node_paths=PackedStringArray("_health", "_hurtbox", "_velocityComponent", "_animation", "_pickupsBox", "_sprite", "_leftWallDetect", "_rightWallDetect", "_miscBox", "_weapon", "_debugHealthLabel") groups=["Players"]]
+[node name="player_elgato" type="CharacterBody2D" node_paths=PackedStringArray("_playerController") groups=["Players"]]
 z_index = 1
 collision_mask = 6
 safe_margin = 2.5
 script = ExtResource("1_q6xoq")
-_health = NodePath("HealthComponent")
-_hurtbox = NodePath("HurtboxComponent")
-_velocityComponent = NodePath("VelocityComponent")
-_animation = NodePath("AnimationComponent")
-_pickupsBox = NodePath("PickupsComponent")
-_sprite = NodePath("sprite")
-_leftWallDetect = NodePath("LeftWallDetect")
-_rightWallDetect = NodePath("RightWallDetect")
-_miscBox = NodePath("MiscBox")
-_weapon = NodePath("Weapon_ElGato")
-_debugHealthLabel = NodePath("HealthLabel")
+_playerController = NodePath("PlayerControllerComponent")
+
+[node name="PlayerControllerComponent" type="Node" parent="." node_paths=PackedStringArray("_health", "_hurtbox", "_velocityComponent", "_animation", "_pickupsBox", "_sprite", "_leftWallDetect", "_rightWallDetect", "_miscBox", "_weapon", "_debugHealthLabel")]
+script = ExtResource("2_40482")
+_health = NodePath("../HealthComponent")
+_hurtbox = NodePath("../HurtboxComponent")
+_velocityComponent = NodePath("../VelocityComponent")
+_animation = NodePath("../AnimationComponent")
+_pickupsBox = NodePath("../PickupsComponent")
+_sprite = NodePath("../sprite")
+_leftWallDetect = NodePath("../LeftWallDetect")
+_rightWallDetect = NodePath("../RightWallDetect")
+_miscBox = NodePath("../MiscBox")
+_weapon = NodePath("../Weapon_ElGato")
+_debugHealthLabel = NodePath("../HealthLabel")
 
 [node name="sprite" type="AnimatedSprite2D" parent="."]
 position = Vector2(0, 3)
