--- Resources/RangedEnemyHeavy.tres
@@ -1,17 +1,0 @@
-[gd_resource type="Resource" script_class="RangedEnemyStats" load_steps=2 format=3 uid="uid://de3fgx3nh7wsq"]
-
-[ext_resource type="Script" path="res://Resources/RangedEnemyStats.cs" id="1_vj2rc"]
-
-[resource]
-script = ExtResource("1_vj2rc")
-RangedEnemyType = 1
-Health = 60
-AttackCooldownTime = 1.5
-BulletDamage = 5
-Knockback = 40.0
-HurtStaggerTime = 1.0
-BulletsPerShot = 5
-BulletAngle = 5.0
-BulletSpeed = 300.0
-BulletDespawnTimeSeconds = 2.5
-RapidFireTime = 0.0
