--- Resources/RangedEnemyLight.tres
@@ -1,17 +1,0 @@
-[gd_resource type="Resource" script_class="RangedEnemyStats" load_steps=2 format=3 uid="uid://dtehwcrv7qseh"]
-
-[ext_resource type="Script" path="res://Resources/RangedEnemyStats.cs" id="1_q6bwc"]
-
-[resource]
-script = ExtResource("1_q6bwc")
-RangedEnemyType = 0
-Health = 20
-AttackCooldownTime = 0.9
-BulletDamage = 5
-Knockback = 25.0
-HurtStaggerTime = 0.9
-BulletsPerShot = 1
-BulletAngle = 2.3
-BulletSpeed = 300.0
-BulletDespawnTimeSeconds = 2.0
-RapidFireTime = 0.0
